+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## Upcoming
 
+- **Improved** error logs in the browser console
+  ([#721](https://github.com/aws/graph-explorer/pull/721))
 - **Updated** dependencies
   ([#718](https://github.com/aws/graph-explorer/pull/718),
   [#720](https://github.com/aws/graph-explorer/pull/720))